Add location syntax to redirection file descriptor
To support parsing the `IO_LOCATION` token that is specified in the
POSIX standard, the `fd` field of the `Redir` struct is changed to
`Option<RedirFd>` instead of `Option<Fd>`. The `RedirFd` enum is added
to represent the file descriptor in redirections.

This commit only changes the abstract syntax tree definition. The
parser does not yet recognize the new syntax.

The `fd_or_default` method of the `Redir` struct is deprecated because
it does not return a meaningful value for the new `RedirFd::Location`
variant. To help finding the default file descriptor for a redirection
operator, the `default_fd` method is added to the `RedirOp` enum.