Add must_use messages
Some functions are already marked with `#[must_use]`, but there were no
messages explaining why. Clippy 0.1.87 is now complaining about this
because of the `yash_env::semantics::Result` value contained in the
return value, which is already marked with `#[must_use]`.
We now add messages to the `#[must_use]` attributes of these functions
to explain that the returned message should also be used as well.