More generic example for if command
The first two examples in the `if` command section used the `[` utility
to check conditions, which might mislead readers into thinking that the
`[` and `]` tokens are part of the `if` command syntax. This commit
replaces those examples with a more generic one that uses `diff -q`,
so that the examples imply that the `if` command conditions can be
arbitrary commands.