docs: Add bg built-in documentation
This commit adds documentation for the built-in. The documentation is
moved from the module documentation for the built-in implementation to
the documentation file for the built-in, removing the leading `//! `
from each line and adding `#` to each heading. The `Implementation
notes` section is left unchanged, as it is not for end users.