Highlight correct source fragment with multi-byte characters
Spans used by annotate-snippets 0.11.x must be byte ranges, but
yash-syntax was providing character ranges as used by previous versions
of annotate-snippets. This caused incorrect highlighting in error
messages when the source contained multi-byte characters, or even
crashes when the character range did not correspond to a valid byte
range.

This commit adds the `Location::byte_range` method to convert a
character range to a byte range, and uses it when converting to
`annotate_snippets::Message`.

Fixes https://github.com/magicant/yash-rs/issues/587