Refactor built-in command execution
The calls to perform_assignments, print, and the built-in execute
function were mostly the same between special built-ins and other
built-ins, so we can combine them. The notable difference is that
special built-ins operate on the current environment, while other
built-ins operate on a pushed volatile context. Previously, this led to
some code duplication. We can use the Either type to handle this
difference in a more elegant way, reducing code duplication and
improving readability.