Reorganize pipefail tests into pipeline-p.sh
We should test the pipefail option in conjunction with the ! keyword.
Move the pipefail tests from option-p.sh to pipeline-p.sh, and add
tests for negated pipelines.

Also add a test to ensure that enabling the pipefail option in a
pipeline does not affect the exit status of that pipeline, as specified
by POSIX-1.2024.