Change Target::Builtin::path non-optional
The path is only used for substitutive built-ins, so it does not provide
any information for other types of built-ins. That is, whether the path
is `None` or `Some("")` does not make any difference for any
functionality. Therefore, we can simplify the code by making this
field non-optional and using an empty string for built-in types other
than `Substitutive`.