Inject prompt dependency into yash-builtin thru yash-env
This is the first step to flatten the dependency graph by removing the
direct dependency of yash-builtin on yash-prompt.
<https://github.com/magicant/yash-rs/issues/621>

The yash-builtin crate no longer depends directly on the yash-prompt
crate. The function to obtain prompt strings is now injected into the
crate via the environment's any storage, using the GetPrompt type
defined in the yash-env crate.