Bump yash-semantics to 0.11.0
Commit 4621573b replaced some modules in yash-semantics with re-exports
from yash-env. This should hardly break anything because all the items
are still available at the same paths, but it might affect users who
implement third-party traits for those types. It's probably safe to
release this as a minor update, but semver-checks is failing because of
<https://github.com/obi1kenobi/cargo-semver-checks/issues/355>. I'm
bumping the version to 0.11.0 as a workaround.