Inject RunFunction to command built-in via Env::any
This is part of the ongoing effort to flatten the dependency graph by
removing yash-semantics as a dependency of yash-builtin. The command
built-in needs to invoke shell functions, which was previously done via
a direct dependency on yash-semantics. Now, the built-in will obtain
the necessary function invocation behavior via an instance of
RunFunction stored in the environment's any storage. This allows the
built-in to remain agnostic of the actual implementation of function
invocation.