Inject external utility execution via Env::any
This is part of the ongoing effort to flatten the dependency graph by
removing direct dependency on yash-semantics from yash-builtin
(<https://github.com/magicant/yash-rs/issues/625>). By injecting the
ability to run external utilities in a subshell via the environment's
`any` storage, we can decouple the `command` built-in from
yash-semantics.