Move error reporting out of replace_current_process
The yash-env crate should not define the text of error messages, as it
is a low-level library. As a preparetion to move the
replace_current_process function to yash-env, this change moves error
reporting out of the function. Now, the function returns an error
value when it fails to execute an external utility, and the caller is
responsible for reporting the error.

This change also allows constructing more detailed error messages, as
the caller has access to more context. The `exec` built-in now uses
`report_failure` for error reporting, which includes the name of the
built-in utility being executed.