Disable yash-semantics feature by default in yash-builtin
The `yash-semantics` feature is now disabled by default in the
`yash-builtin` crate. To enable it, you must explicitly specify the
feature in your `Cargo.toml` file.

The `Cargo.toml` file for the `yash3` crate is not modified, so the
`yash-semantics` feature is no longer enabled when building the `yash3`
binary. This allows `yash-builtin` to be built before `yash-semantics`
is built, improving parallelism in the build process.

Closes https://github.com/magicant/yash-rs/issues/625