Remove versions from dev-dependencies in workspace crates
Specifying `workspace = true` means that the dependency has to be
released before the dependant crate can be released. This was causing
bootstrapping issues because of circular dependencies between crates in
the workspace. To resolve this, dev-dependencies are now specified
directly via path, which do not impose release order constraints. Note
that this only affects dev-dependencies, as normal dependencies still
need to require the correct version of the dependency.