Make yash-env unit tests pass
This commit applies a series of changes to the test code in the yash-env
crate to ensure compatibility with recent refactorings. The primary
changes include updating the Env type to be generic over the system
type, removing unnecessary boxing of system instances, and adjusting
the FunctionBody trait implementations to accommodate the generic Env.